requirements.txt
python_barcode==0.14.0
codetiming==1.4.0
Faker==18.5.1
imgaug==0.4.0
importlib_metadata==4.6.4
matplotlib==3.6.3
numpy==1.24.1
#numpy==1.21.5
opencv_python==4.7.0.68
#Pillow==9.0.1
Pillow==9.5.0
qrcode==7.4.2
resize_image==0.4.0
seaborn==0.12.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.5.0
wandb==0.15.0
pytest==7.3.1
pytest-asyncio==0.21.0